requirements.txt
absl-py==2.3.0
amqp==5.3.1
appnope==0.1.4
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
cffi==1.17.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
Django==5.2.3
djangorestframework==3.16.0
executing==2.2.0
flatbuffers==25.2.10
fonttools==4.58.4
gunicorn==23.0.0
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jax==0.6.2
jaxlib==0.6.2
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
kombu==5.5.4
matplotlib==3.10.3
matplotlib-inline==0.1.7
mediapipe==0.10.21
ml_dtypes==0.5.1
nest-asyncio==1.6.0
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
protobuf==4.25.8
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycairo==1.28.0
pycparser==2.22
Pygments==2.19.2
PyGObject==3.52.3
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pyzmq==27.0.0
redis==6.2.0
scipy==1.15.3
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.2
sqlparse==0.5.3
stack-data==0.6.3
supervisor==4.2.5
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.14.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
webrtcvad==2.0.10
